fix#246: 결과페이지 응답 스팩 리스트로 수정
