+++ src/main/java/or/sopt/houme/domain/user/controller/dto/ImageHistoriesResultPageResponse.java
@@ -1,0 +1,33 @@
+package or.sopt.houme.domain.user.controller.dto;
+
+import java.util.List;
+
+public record ImageHistoriesResultPageResponse(
+        List<ImageHistoryResultPageResponse> histories
+) {
+    public static ImageHistoriesResultPageResponse of(List<ImageHistoryResultPageResponse> histories) {
+        return new ImageHistoriesResultPageResponse(histories);
+    }
+
+    public record ImageHistoryResultPageResponse(
+            String equilibrium,
+            String houseForm,
+            String tasteTag,
+            String name,
+            String generatedImageUrl,
+            boolean isLike
+    ) {
+        public static ImageHistoryResultPageResponse of(
+                String equilibrium,
+                String houseForm,
+                String tasteTag,
+                String name,
+                String generatedImageUrl,
+                boolean isLike
+        ) {
+            return new ImageHistoryResultPageResponse(
+                    equilibrium, houseForm, tasteTag, name, generatedImageUrl, isLike
+            );
+        }
+    }
+}
